Added AddExperience logic. Level is now float for characters.

+++ Source/May/Public/AbilitySystem/Utils/MayAbilitySystemLibrary.h
@@ -20,10 +20,10 @@ class MAY_API UMayAbilitySystemLibrary : public UBlueprintFunctionLibrary {
 
 public:
 	UFUNCTION(BlueprintCallable, Category="MayAbilitySystemLibrary")
-	static void InitEnemyDefaultAttributes(const UObject* WorldContextObject, UAbilitySystemComponent* ASC, const EEnemyType EnemyType, const int32 Level);
+	static void InitEnemyDefaultAttributes(const UObject* WorldContextObject, UAbilitySystemComponent* ASC, const EEnemyType EnemyType, const float Level);
 	
 	UFUNCTION(BlueprintCallable, Category="MayAbilitySystemLibrary")
-	static void InitEnemyDefaultAbilities(const UObject* WorldContextObject, UAbilitySystemComponent* ASC, const EEnemyType EnemyType, const int32 Level);
+	static void InitEnemyDefaultAbilities(const UObject* WorldContextObject, UAbilitySystemComponent* ASC, const EEnemyType EnemyType, const float Level);
 
 	UFUNCTION(BlueprintCallable, Category="MayAbilitySystemLibrary")
 	static FEnemyTypeDefaultInfo GetEnemyTypeInfo(const UObject* WorldContextObject, const EEnemyType EnemyType);
